refactor(browser-type-check): convert files to typescript
