--- src/__internal__/utils/helpers/browser-type-check/browser-type-check.js
@@ -1,15 +1,0 @@
-/**
- * Check for Chrome or Firefox.
- */
-function browserTypeCheck(_window) {
-  return Boolean(_window.chrome || _window.sidebar);
-}
-
-export function isSafari(navigator) {
-  return navigator.vendor.includes("Apple");
-}
-
-export function isEdge(navigator) {
-  return navigator.userAgent.includes("Edge");
-}
-export default browserTypeCheck;
