--- src/__internal__/utils/helpers/browser-type-check/index.js
@@ -1,1 +1,0 @@
-export { default, isSafari, isEdge } from "./browser-type-check";
